Move history tests into test_release_artifacts.py and drop test_protocol_history.py
